+++ src/piidigger/globalfuncs.py
@@ -18,7 +18,7 @@ from piidigger.globalvars import maxChunkSize
 #   ext: [a list of file extensions with the leading .]
 #   mime: [a list of mime-type strings]
 
-fileHandlers={ handler: getattr(getattr(fh, handler), 'handles') for handler in fh.__dir__() if not handler.startswith('__') }
+fileHandlers={ handler: getattr(getattr(fh, handler), 'handles') for handler in fh.__dir__() if not handler.startswith('_') }
 
 # Leave these in here as reference until I write more file handlers.
 # fileHandlers={
@@ -72,28 +72,6 @@ fileHandlers={ handler: getattr(getattr(fh, handler), 'handles') for handler in 
 ############### Global functions #####################
 ######################################################
 
-
-def appendContent(content: str, line: str, maxContentSize: int):
-    '''Appends a line of text to a content string, returning the updated content string'''
-
-    if (len(content.strip()) + len(line.strip())) < maxContentSize:
-        if content == '':
-            content = line.replace('\t', ' ').strip()
-        else:
-            content += ' ' + line.replace('\t', ' ').strip()
-    else:
-        i: int = 0
-        words: list = line.split()
-        for word in words:
-            i+=1
-            if content == '':
-                content = word
-            else:
-                content += ' ' + word
-            if len(content.strip()) > maxContentSize:
-                return (content.strip(), ' '.join(words[i:])) 
-    
-    return content.strip(), ''
 
 def clearQ(q: Queue):
     '''Clears a queue of all contents'''
